requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
blinker==1.6.2
build==0.10.0
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.22
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickhouse-connect==0.5.24
dataclasses-json==0.5.7
faiss-cpu==1.7.4
fastapi==0.95.1
filelock==3.12.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
frozenlist==1.3.3
fsspec==2023.5.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
langchain==0.0.219
langchainplus-sdk==0.0.17
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
posthog==3.0.1
prompt-toolkit==3.0.38
protobuf==4.23.0
psycopg2-binary==2.9.6
ptvsd==4.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==1.10.7
PyJWT==2.8.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.3.1
pytest-flask==1.2.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
starlette==0.26.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
todoist-api-python==2.0.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
waitress==2.1.2
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.3
wikipedia==1.4.0
yarl==1.9.2
youtube-transcript-api==0.6.0
zipp==3.15.0
zstandard==0.21.0
